Moving one last test to flakey runner
